tidying and various refactors

+++ arek_chess/criteria/selection/fast_selector.py
@@ -3,10 +3,12 @@
 Selecting one best node fast.
 """
 
+from numpy import double
+
 from typing import List
 
+from arek_chess.common.constants import INF
 from arek_chess.criteria.selection.base_selector import BaseSelector
-from arek_chess.constants import INF
 from arek_chess.main.game_tree.node.node import Node
 
 
@@ -17,7 +19,7 @@ class FastSelector(BaseSelector):
 
     def select(self, nodes: List[Node], color: bool) -> Node:
         best_node: Node
-        best_score: float = -INF if color else INF
+        best_score: double = -INF if color else INF
 
         for node in nodes:
             # node white-to-move then best child is one with the highest score
